Refactored database model so items are in their own file
