+++ package.json
@@ -58,6 +58,7 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.14.3",
+    "faker": "^4.1.0",
     "file-loader": "^4.1.0",
     "jest": "^24.8.0",
     "jest-cli": "^24.8.0",
